[FIX] Add missing static_cast for scoped enum array indexing
Scoped enums don't implicitly convert to integers, so explicit
static_cast<size_t>() is needed when using them as array indices
in the visualizer classes.

+++ src/openms_gui/source/VISUAL/VISUALIZER/InstrumentSettingsVisualizer.cpp
@@ -38,7 +38,7 @@ namespace OpenMS
     if (!isEditable())
     {
       fillComboBox_(instrumentsettings_scan_mode_, &temp_.NamesOfScanMode[temp_.getScanMode()], 1);
-      fillComboBox_(instrumentsettings_polarity_, &IonSource::NamesOfPolarity[temp_.getPolarity()], 1);
+      fillComboBox_(instrumentsettings_polarity_, &IonSource::NamesOfPolarity[static_cast<size_t>(temp_.getPolarity())], 1);
 
     }
     else
